Using Webhooks to give feedback from emails to database
we use sendgrid to sending an email, using webhook to help snedgrid communicate with our server and save respond data from email to mongodb. In devlopment, using Ngrok as a tunnel to connect sendgrid with our api. filtering the data from webhooks and make a good efficient query to db